scroll fil actu ajax done

+++ post_filactu.php
@@ -1,20 +1,34 @@
- <?php include('connexionbdd.php');?>
-
+<?php include('connexionbdd.php');?>
+<?php session_start(); ?>
 
 <?php 
-$file_filactu = $bdd->prepare('SELECT pseudo,message,new_name_file FROM fichier_upload ORDER BY date_upload DESC LIMIT 0, 3');
+$i=$_POST["var"];
+
+$compteur = $bdd->prepare('SELECT COUNT(id) FROM fichier_upload');
+$compteur->execute();
+$array_post=$compteur->fetch();
+$compteur_post=$array_post[0];
+setcookie("compteur_post",$compteur_post);
+$file_filactu = $bdd->prepare('SELECT pseudo,message,new_name_file FROM fichier_upload ORDER BY date_upload DESC LIMIT '.$i.', 5');
 $file_filactu->execute();
+
+
 $extimage=array(".png",".jpg",".jepg",".gif",".PNG",".JPG",".JPEG",".GIF");
 $extvideo=array(".mp4",".avi",".mov",".mpg",".mpeg",".mvw",".MP4",".AVI",".MOV",".MPG",".MPEG",".MVW");
 $extaudio=array(".mp3",".MP3");
+    
+
     while($file_exist=$file_filactu->fetch()){
-        echo "<div class='postfilactu'>";
         
+        echo "<div class='postfilactu'>";
+        if ($_SESSION['pseudo']==$file_exist['pseudo']){
         echo "<div class='modifsupp'> <button>modifier</button><button>suprimer</button> </div> ";
+        }
+        
         echo " <div> <strong> " .htmlspecialchars($file_exist['pseudo']). "</p></strong> </div>    ";
         if(isset($file_exist["message"])){
         echo " <div> <p>" .htmlspecialchars($file_exist["message"]). "</p> </div>  ";
-    }
+        }
         if(in_array(strrchr($file_exist["new_name_file"], "."),$extimage)){
         echo "<img src='upload/" .$file_exist["new_name_file"]."' />";
         }
